requirements.txt
numpy==1.26.3
pillow==11.3.0
